requirements.txt
acconeer-exptool==7.11.1
amqp==5.2.0
asgiref==3.8.1
attributes-doc==0.3.0
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
cbor2==5.6.4
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
commonmark==0.9.1
constantly==23.10.4
contourpy==1.2.1
cron-descriptor==1.4.3
cryptography==42.0.8
cycler==0.12.1
daphne==4.1.2
Django==5.0.6
django-celery-beat==2.6.0
django-cors-headers==4.4.0
django-redis==5.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
drf-yasg==1.21.7
exceptiongroup==1.1.1
fonttools==4.53.1
h5py==3.11.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
kiwisolver==1.4.5
kombu==5.3.7
matplotlib==3.9.2
msgpack==1.0.8
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pip==24.1.2
prompt_toolkit==3.0.47
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.18.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
pyserial==3.5
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
pyusb==1.2.1
PyYAML==6.0.1
redis==5.0.7
requests==2.32.3
result==0.16.1
rich==12.6.0
scipy==1.14.0
service-identity==24.1.0
setuptools==70.1.0
six==1.16.0
sqlparse==0.5.0
textual==0.1.18
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
XlsxWriter==3.2.0
zope.interface==6.4.post2
